refactor(setting): 워크스페이스 DTO 구조 정리
