+++ src/workspaces/application/dtos/my-subscription-output.dto.ts
@@ -1,0 +1,11 @@
+import type {
+  WorkspaceSubscriptionPlan,
+  WorkspaceSubscriptionStatus,
+} from '../../domain/workspace.types';
+
+export type MySubscriptionOutput = {
+  plan: WorkspaceSubscriptionPlan;
+  status: WorkspaceSubscriptionStatus;
+  autoRenew: boolean;
+  currentPeriodEnd: Date | null;
+};
