+++ src/workspaces/application/dtos/update-my-subscription-input.dto.ts
@@ -1,0 +1,9 @@
+import type {
+  WorkspaceSubscriptionAction,
+  WorkspaceSubscriptionPlan,
+} from '../../domain/workspace.types';
+
+export type UpdateMySubscriptionInput = {
+  type: WorkspaceSubscriptionAction;
+  plan?: WorkspaceSubscriptionPlan;
+};
