+++ src/workspaces/application/usecases/get-my-subscription.usecase.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { WORKSPACES_REPOSITORY } from '../../domain/workspaces.repository';
 import type { WorkspacesRepository } from '../../domain/workspaces.repository';
-import { MySubscriptionResponse } from '../../domain/workspace.types';
+import { MySubscriptionOutput } from '../dtos/my-subscription-output.dto';
 import { normalizeExpiredSubscription } from '../policies/subscription-expiration.policy';
 import { toMySubscriptionResponse } from '../policies/subscription-response.policy';
 
@@ -12,7 +12,7 @@ export class GetMySubscriptionUseCase {
     private readonly workspacesRepository: WorkspacesRepository,
   ) {}
 
-  async execute(userId: string): Promise<MySubscriptionResponse> {
+  async execute(userId: string): Promise<MySubscriptionOutput> {
     const currentSubscription =
       await this.workspacesRepository.getOrCreatePersonalWorkspaceSubscription(
         userId,
